Move TestUpNavigationHandler to testing-unit-common
