+++ core/src/test/java/com/bumble/appyx/core/routing/upnavigation/UpNavigationTest.kt
@@ -12,8 +12,9 @@ import com.bumble.appyx.core.plugin.StubUpNavigationHandler
 import com.bumble.appyx.core.plugin.UpNavigationHandler
 import com.bumble.appyx.routingsource.backstack.BackStack
 import com.bumble.appyx.routingsource.backstack.operation.push
-import com.bumble.appyx.testing.unit.common.util.TestIntegrationPoint
 import com.bumble.appyx.testing.junit4.util.MainDispatcherRule
+import com.bumble.appyx.testing.unit.common.util.TestIntegrationPoint
+import com.bumble.appyx.testing.unit.common.util.TestUpNavigationHandler
 import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
